file_util: move path handling functions from ui module
We'll need this kind of functions in the library crate to parse user file
patterns.
